+++ changelog.md
@@ -2,3 +2,5 @@ Changelog version 1.6.9.7
  
 Renamed crm_configurations.php to enviroment.php
 Moved constant passepartout for login into enviroment.php
+Translated enviroment.php
+Bug fixes